refactor HTML to Air Tags
