+++ src/air/__init__.py
@@ -355,6 +355,6 @@ from .tags import (
 from .tags import (
     Wbr as Wbr,
 )
-from .tags import html_to_airtags as html_to_airtags
 from .templates import JinjaRenderer as JinjaRenderer
 from .templates import Jinja2Renderer as Jinja2Renderer
+from .utils import html_to_airtags as html_to_airtags
